+++ src/gui/mod.rs
@@ -1,5 +1,5 @@
 mod shell;
-mod plot_demo;
+mod views;
 mod screens;
 
 pub use shell::Shell;
